requirements.txt
requests
googletrans
python-telegram-bot